feat: creates root and curriculum layout
